python/micro-services/requirements.txt
redis==3.3.8
celery~=4.4.7
requests==2.28.1
supervisor==4.1.0
fastapi==0.68.0
fastapi-auth0==0.2.1
motor==2.5.0
PyJWT==1.7.1
pymongo==3.12.0
python-jose==3.3.0
python-json-logger==2.0.2
six==1.14.0
uvicorn==0.11.3
paramiko~=2.8.0
python-keycloak-client==0.2.3
boto3==1.19.12
razorpay==1.3.0
anyio==3.5.0
pytest-asyncio==0.18.3
pytest-tornasync==0.6.0.post2
pytest-trio==0.7.0
python-redis-lock==4.0.0
pydenticon==0.3.1
reportlab==4.0.8